moved ArrayBase into Runtime
