+++ src/genn/backends/single_threaded_cpu/backend.cc
@@ -1405,8 +1405,8 @@ void Backend::genStepTimeFinalisePreamble(CodeStream &, const ModelSpecMerged &)
 {
 }
 //--------------------------------------------------------------------------
-std::unique_ptr<ArrayBase> Backend::createArray(const Type::ResolvedType &type, size_t count, 
-                                                VarLocation location, bool uninitialized) const
+std::unique_ptr<Runtime::ArrayBase> Backend::createArray(const Type::ResolvedType &type, size_t count, 
+                                                         VarLocation location, bool uninitialized) const
 {
     return std::make_unique<Array>(type, count, location, uninitialized);
 }
